Clean

+++ Main/Events/DrinkEvent/DrinkButtonClickEvent.cs
@@ -22,7 +22,7 @@ namespace Main.Events.DrinkEvent
                     MessageBox.Show("Please fill in complete information.", Reference.Warning, MessageBoxButtons.OK, MessageBoxIcon.Warning);
                     return;
                 }
-                else if (!int.TryParse(d.tb_drink_id.Text, out int result) || !int.TryParse(d.tb_price.Text, out int result1))
+                else if (!int.TryParse(d.tb_drink_id.Text, out int result) || !double.TryParse(d.tb_price.Text, out double result1))
                 {
                     MessageBox.Show("Drink ID and price must be numbers only.", Reference.Warning, MessageBoxButtons.OK, MessageBoxIcon.Warning);
                     return;
